services/files/pyproject.toml
[tool.poetry]
name = "files"
version = "0.1.0"
description = "File storage service"
authors = ["ilia <iliaforwork1993@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11"
fastapi = "^0.111.1"
uvicorn = {extras = ["standard"], version = "^0.30.3"}
sqlalchemy = "^2.0.31"
asyncpg = "^0.29.0"
alembic = "^1.13.2"
boto3 = "^1.34.143"
python-multipart = "^0.0.9"
python-dotenv = "^1.0.1"
kafka-python = "^2.0.2"
pyjwt = {extras = ["crypto"], version = "^2.8.0"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
